1. hide/show homefee 2. add mixins user 3. change feecommon to script fordel 4. fixbug router sign_up_done alway permission to access
