+++ src/scripts/feeCommon.js
